Contract: Bridge fees collection and claiming implementation (#60)
- Implemented fee collection and claiming for all tokens. 
How it works: Every time a user sends a token it will collect fees into
the contract (part of the amount). The truncated amount counts towards
the fee (so that users can choose what they want to receive on the other
side).
For claiming, any relayer can claim when ever he wants and the fees will
be proportionally distributed to all of them.